+++ Held-Karp-algorithm/Held-Karp-algorithm/HeldKarp.hpp
@@ -27,24 +27,24 @@ namespace TSP
 		struct sInfo
 		{
 			uint_least16_t cost;
-			vector<uint_least8_t> path;
+			vector<uint_least16_t> path;
 		};
 
 		// <Code, <Node, Data>>
-		queue<map<uint_least32_t, map<uint_least8_t, sInfo>>> C;
+		queue<map<uint_least32_t, map<uint_least16_t, sInfo>>> C;
 
 	protected:
 		void AddNewToQueue();
 
-		string PrintPath(const uint_least32_t code, const uint_least8_t π);
+		string PrintPath(const uint_least32_t code, const uint_least16_t π);
 
-		void Combinations(const uint_least8_t K, const uint_least8_t N);
-		void Combinations_FreeMem(stack<uint_least8_t> &Q, vector<uint_least8_t> &S, const uint_least8_t K, const uint_least8_t N, const uint_least8_t sCur);
+		void Combinations(const uint_least16_t K, const uint_least16_t N);
+		void Combinations_FreeMem(stack<uint_least16_t> &Q, vector<uint_least16_t> &S, const uint_least16_t K, const uint_least16_t N, const uint_least16_t sCur);
 
 		void Solve(uint_least16_t &opt, string &path);
 
 	public:
-		HeldKarp(const vector<vector<uint_least8_t>> &DistanceMatrix2D);
+		HeldKarp(const vector<vector<uint_least16_t>> &DistanceMatrix2D);
 
 	};
 }
