+++ Held-Karp-algorithm/Held-Karp-algorithm/TSP.hpp
@@ -26,19 +26,19 @@ namespace TSP
 		// precalculated
 		uint_least32_t POWER2[31] = { 1, 2, 4, 8, 16, 32, 64, 128, 256, 512, 1024, 2048, 4096, 8192, 16384, 32768, 65536, 131072, 262144, 524288, 1048576, 2097152, 4194304, 8388608, 16777216, 33554432, 67108864, 134217728, 268435456, 536870912, 1073741824 };
 
-		const vector<vector<uint_least8_t>> distance;
-		const uint_least8_t numberOfNodes;
+		const vector<vector<uint_least16_t>> distance;
+		const uint_least16_t numberOfNodes;
 
 		time_point<system_clock> begin;
-		atomic<uint_least8_t> currentCardinality = 2;
+		atomic<uint_least16_t> currentCardinality = 2;
 		atomic<bool> writingBuffer = false;
 
 	protected:
-		uint_least32_t Powered2Code(const vector<uint_least8_t> &S);
+		uint_least32_t Powered2Code(const vector<uint_least16_t> &S);
 
-		uint_least32_t Powered2Code(const vector<uint_least8_t> &S, const uint_least8_t exclude);
+		uint_least32_t Powered2Code(const vector<uint_least16_t> &S, const uint_least16_t exclude);
 
-		uint_least32_t Powered2Code(const uint_least32_t code, const uint_least8_t exclude);
+		uint_least32_t Powered2Code(const uint_least32_t code, const uint_least16_t exclude);
 
 		void ETL();
 		void ETLw();
@@ -49,11 +49,11 @@ namespace TSP
 		virtual void Solve(uint_least16_t &opt, string &path) = 0;
 
 	public:
-		TSP(const vector<vector<uint_least8_t>> &DistanceMatrix2D);
+		TSP(const vector<vector<uint_least16_t>> &DistanceMatrix2D);
 
 		void Run();
 
-		static vector<vector<uint_least8_t>> New_RND_Distances(const uint_least8_t Size_of_RandomDistanceCosts);
+		static vector<vector<uint_least16_t>> New_RND_Distances(const uint_least16_t Size_of_RandomDistanceCosts);
 
 	};
 }
